feat: move browsers to configuration.nix file
When trying to share the screen using either Firefox or Vivaldi
a GNOME pop-up with selections is shown but nothing really works
and the screen is always black for other participants.

Apparently this is fixed by moving the browser pkg definition
out of home-manager into the system configuration.nix, which
for my case, actually solved my issue.